requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cryptography==43.0.3
cymem==2.0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
fastapi==0.115.5
fpdf==1.7.2
h11==0.14.0
idna==3.10
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
language_tool_python==2.8.1
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.10
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
preshed==3.0.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.17
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.41.2
thinc==8.3.2
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
PyMuPDF